Override old kernel naming
